pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1"
authors = ["Huw Cheston <h.cheston@qmul.ac.uk>", "Iran Roman <i.roman@qmul.ac.uk>"]
license = "CC-BY-4.0"
readme = "README.md"
packages = [
	{ include = "mirexplorer_common", from = "src" },
	{ include = "mirexplorer_audio", from = "src" },
	{ include = "mirexplorer_analysis", from = "src" },
	{ include = "mirexplorer_chat", from = "src" },
	{ include = "mirexplorer_gateway", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = {extras = ["async"], version = "^3.1.2"}
Flask-Caching = "*"
flask-wtf = "*"
librosa = "*"
apscheduler = "*"
loguru = "*"
pydub = "*"
python-dotenv = "*"
yarl = "*"
pydantic-ai = "*"
flask-limiter = "^4.1.1"
gunicorn = "^23.0.0"
mutagen = "^1.47.0"
setuptools = "^80.10.2"
google-cloud-storage = "^3.8.0"
httpx = ">=0.27"
requests = ">=2.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"